Document and reorganize the FCB gen crate
